+++ tests/unit/multi_table/test_base.py
@@ -40,7 +40,9 @@ class TestBaseMultiTableSynthesizer:
         locales = ['en_CA', 'fr_CA']
         instance = Mock()
         instance._table_synthesizers = {}
-        instance._table_parameters = {'nesreca': {'default_distribution': 'gamma'}}
+        instance._table_parameters = {
+            'nesreca': {'default_distribution': 'gamma', 'locales': ['en_US']},
+        }
         instance.locales = locales
         instance.metadata = get_multi_table_metadata()
 
@@ -57,7 +59,7 @@ class TestBaseMultiTableSynthesizer:
             call(
                 metadata=instance.metadata.tables['nesreca'],
                 default_distribution='gamma',
-                locales=locales,
+                locales=['en_US'],
             ),
             call(metadata=instance.metadata.tables['oseba'], locales=locales),
             call(metadata=instance.metadata.tables['upravna_enota'], locales=locales),
@@ -763,6 +765,7 @@ class TestBaseMultiTableSynthesizer:
                 'id_upravna_enota': np.arange(10),
             }),
         }
+        instance._transform_helper = Mock(return_value=data)
 
         synth_nesreca = Mock()
         synth_oseba = Mock()
@@ -782,6 +785,7 @@ class TestBaseMultiTableSynthesizer:
             'oseba': synth_oseba._preprocess.return_value,
             'upravna_enota': synth_upravna_enota._preprocess.return_value,
         }
+        instance._transform_helper.assert_called_once_with(data)
         instance.validate.assert_called_once_with(data)
         assert instance.metadata._get_all_foreign_keys.call_args_list == [
             call('nesreca'),
@@ -1212,6 +1216,7 @@ class TestBaseMultiTableSynthesizer:
             'table2': pd.DataFrame({'id': [1, 2, 3], 'name': ['John', 'Johanna', 'Doe']}),
         }
         instance._sample = Mock(return_value=data)
+        instance._reverse_transform_helper = Mock(return_value=data)
 
         synth_id = 'BaseMultiTableSynthesizer_1.0.0_92aff11e9a5649d1a280990d1231a5f5'
         instance._synthesizer_id = synth_id
@@ -1222,6 +1227,7 @@ class TestBaseMultiTableSynthesizer:
 
         # Assert
         instance._sample.assert_called_once_with(scale=1.5)
+        instance._reverse_transform_helper.assert_called_once_with(data)
         assert caplog.messages[0] == str({
             'EVENT': 'Sample',
             'TIMESTAMP': '2024-04-19 16:20:10.037183',